AE-1718: Move Fieldset component to its own folder
